Enhance Parameter Tuning and Genetic Algorithm Runner with Improved Configuration and Error Handling
- Updated ParameterTuningManager to return results and improve CSV output
- Simplified ParameterParser grid generation with new ParameterSet methods
- Refactored GeneticAlgorithmRunner with streamlined component creation
- Added new methods to GeneticAlgorithm for process ID and best solution tracking
- Improved error handling and logging in genetic algorithm execution
- Updated CMakeLists.txt to include new tuning and metaheuristics source files

+++ include/metaheuristics/selection_impl.h
@@ -35,10 +35,10 @@ public:
 
         try {
             std::vector<std::shared_ptr<Individual>> selected;
-            selected.reserve(m_config.getParentCount());
+            selected.reserve(static_cast<size_t>(m_config.getParentCount()));
 
             auto& rng = Random::instance();
-            while (selected.size() < m_config.getParentCount()) {
+            while (selected.size() < static_cast<size_t>(m_config.getParentCount())) {
                 std::vector<std::shared_ptr<Individual>> tournament;
                 tournament.reserve(m_config.getTournamentSize());
